% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneDeMix.r
\name{CloneDeMix}
\alias{CloneDeMix}
\title{Decomposing the subclonal structure of tumors with two-way mixture models on copy number aberrations.}
\usage{
CloneDeMix(tumor, normal = NULL, threshold = 10^-5, iterC = 10^3,
  CNVstate = c(0:10), method = "aic")
}
\arguments{
\item{tumor}{A G by N expression matrix of tumor samples, where G is the gene number and N is the sample size.}

\item{normal}{A vector of length s. It's expression profile from the paired normal sample. If it's empty, the baseline is calculated from sample mean.}

\item{threshold}{The threhold of convergence. Default is 10^-5.}

\item{CNVstate}{A vector of considered copy number state.}

\item{method}{The criteria for the final clone number.}

\item{iterationC}{Maximum number of iterations. Default is 1000.}
}
\value{
A list is containing an estimated CNV matrix, MCP matrix, and the number of predicted clone number.
}
\description{
A R package for deconvoluting subclonal copy number and mutational cellular prevalence.
}
\examples{
data("ESCC_chr1")
res <- CloneDeMix(tumor=ESCC_chr1$tumor, normal=ESCC_chr1$normal)
head(res$CNV); head(res$MCP)
}
\author{
An-Shun Tai \email{daansh13@gmail.com}
}
