% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneDeMix_core.R
\name{CloneDeMix_core}
\alias{CloneDeMix_core}
\title{Decomposing the subclonal structure of tumors with two-way mixture models on copy number aberrations.}
\usage{
CloneDeMix_core(dataX, threshold = 10^-5, CNVstate = c(0:4), numbG = 4,
  baseline = NULL, iterationC = 10^3)
}
\arguments{
\item{dataX}{Gene expression of a tumor sample.}

\item{threshold}{The threhold of convergence. Default is 10^-5.}

\item{CNVstate}{A vector of considered copy number state.}

\item{numbG}{The number of subclones.}

\item{baseline}{A vector of length s. It's expression profile from the paired normal sample. If it's empty, the baseline is calculated from sample mean.}

\item{iterationC}{Maximum number of iterations. Default is 1000.}
}
\value{
A list containing MCP vector, the weight of copy number state, error rate, AIC, and BIC.
}
\description{
The core function of CloneDeMix.
}
\author{
An-Shun Tai \email{daansh13@gmail.com}
}
